+++ components/homeIndex/pageDetail.tsx
@@ -202,8 +202,14 @@ const PageDetail = (props: { data: IDemographicInsight; items: IInstagramerHomeT
         </div>
         <div className={`${styles.frameContainer} ${isHidden ? "" : styles.show}`}>
           <section className={styles.pageDetailSection}>
-            <div className="title2" style={{ textAlign: "center" }}>
-              {t(LanguageKey.AudienceGender)} ({t(LanguageKey.home_Last30Days)})
+            <div className="title2">
+              {t(LanguageKey.AudienceGender)}{" "}
+              <Tooltip
+                tooltipValue={t(LanguageKey.AudienceTooltip)}
+                triggerType="tooltip"
+                onClick={true}
+                position="bottom"
+              />
             </div>
             <div className="explain">
               {t(LanguageKey.AudienceGenderStats, {
@@ -253,8 +259,15 @@ const PageDetail = (props: { data: IDemographicInsight; items: IInstagramerHomeT
 
           <section className={styles.pageDetailSection}>
             <div className="title2">
-              {t(LanguageKey.AudienceAge)} ({t(LanguageKey.home_Last30Days)})
+              {t(LanguageKey.AudienceAge)}{" "}
+              <Tooltip
+                tooltipValue={t(LanguageKey.AudienceTooltip)}
+                triggerType="tooltip"
+                onClick={true}
+                position="bottom"
+              />
             </div>
+
             <div className="explain">
               {(() => {
                 // Find the age range with highest percentage
@@ -390,7 +403,13 @@ const PageDetail = (props: { data: IDemographicInsight; items: IInstagramerHomeT
           <section className={styles.pageDetailSection}>
             <div className="headerandinput">
               <div className="title2">
-                {t(LanguageKey.AudienceLocation)} ({t(LanguageKey.home_Last30Days)})
+                {t(LanguageKey.AudienceLocation)}{" "}
+                <Tooltip
+                  tooltipValue={t(LanguageKey.AudienceTooltip)}
+                  triggerType="tooltip"
+                  onClick={true}
+                  position="bottom"
+                />
               </div>
 
               <div className="explain">